DAL migration

+++ src/BirdsiteLive.Common/Interfaces/SocialMediaUserDal.cs
@@ -32,7 +32,7 @@ public interface SyncUser
 
         public int FetchingErrorCount { get; set; } //TODO: update DAL
         public long Followers { get; set; } 
-        public long StatusesCount { get; set; }
+        public int StatusesCount { get; set; }
         public JsonElement ExtraData { get; set; }
         public string PreDescriptionHook { get; }
         public string PostDescriptionHook { get; }
